requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
pillow==10.0.0
flask==2.3.3
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
opencv-python==4.8.0.76
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
h5py==3.9.0 